% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rp_connect}
\alias{rp_connect}
\title{Connect to RP database}
\usage{
rp_connect(
  rp_odbc_name = Sys.getenv("RP_ODBC_NAME"),
  database = Sys.getenv("RP_DATABASE"),
  TZ = Sys.timezone(),
  conn = NULL
)
}
\arguments{
\item{rp_odbc_name}{Name of the Windows ODBC data source to use. Retrieved
from \code{RP_ODBC_NAME} environment variable by default but can be overridden here,
see details.}

\item{database}{Name of the default database to connect to. Retrieved
from \code{RP_DATABASE} environment variable by default but can be overridden here,
see details.}

\item{TZ}{The timezone used by the database server.}

\item{conn}{Existing DBI connection to use. If \code{NULL} (the default) a new connection will be opened.}
}
\description{
Establish a connection to a Versatrans Routing and Planning database using
the DBI/odbc packages in R.
}
\details{
You must set up a ODBC data source for the Versatrans RP database. You must
be connected to the BPS network (e.g., via VPN if you are remote). See
https://support.microsoft.com/en-us/office/administer-odbc-data-sources-b19f856b-5b9b-48c9-8b93-07484bfab5a7

#' Connection parameters \code{rp_odbc_name} and \code{database} should usually be
specified in a .Renviron file, see
https://support.posit.co/hc/en-us/articles/360047157094-Managing-R-with-Rprofile-Renviron-Rprofile-site-Renviron-site-rsession-conf-and-repos-conf
if you are not familiar with seting envionment variables in R.
}
\examples{
\dontshow{if (Sys.getenv("RP_ODBC_NAME")!=""&Sys.getenv("RP_DATABASE")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

conn <- rp_connect()
{tbls <- DBI::dbListTables(conn); head(tbls)}
DBI::dbDisconnect(conn)
\dontshow{\}) # examplesIf}
}
