% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitors_student_file.R
\name{rp_report_student_monitor}
\alias{rp_report_student_monitor}
\title{Create student monitor report}
\usage{
rp_report_student_monitor(
  database = Sys.getenv("RP_DATABASE"),
  format = c("long", "wide")
)
}
\arguments{
\item{database}{Character vector of length one specifying the database to connect to}

\item{format}{Character vector of length one specifying how the data should be
formatted. If \code{"long"} (the default) return separate \emph{rows} for each weekday/pickup/dropoff.
If \code{"wide"} return separate \emph{columns} for each weekday/pickup/dropoff.}
}
\description{
Equivalent to \emph{Import/Export => Student => Monitors student file}
}
\examples{
\dontshow{if (Sys.getenv("RP_ODBC_NAME")!=""&Sys.getenv("RP_DATABASE")!="") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

student_monitor <- rp_report_student_monitor()

## data fields:
dplyr::glimpse(student_monitor[0, ])

\dontshow{\}) # examplesIf}
}
