% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_overrides.R
\name{rp_report_vehicle_overrides}
\alias{rp_report_vehicle_overrides}
\title{Get vehicle override report}
\usage{
rp_report_vehicle_overrides(
  rp_odbc_name = Sys.getenv("RP_ODBC_NAME"),
  database = "Onscreen021919",
  TZ = Sys.timezone(),
  conn = NULL
)
}
\arguments{
\item{rp_odbc_name}{Name of the Windows ODBC data source to use. Retrieved
from \code{RP_ODBC_NAME} environment variable by default but can be overridden here,
see details.}

\item{database}{Name of the default database to connect to. Retrieved
from \code{RP_DATABASE} environment variable by default but can be overridden here,
see details.}

\item{TZ}{The timezone used by the database server.}

\item{conn}{Existing DBI connection to use. If \code{NULL} (the default) a new connection will be opened.}
}
\description{
Get vehicle override report
}
